Struct fim::editor::Editor

pub struct Editor<'a> { /* fields omitted */ }

Struct that represents the fim editor.

Implementations

impl<'a> Editor<'a>

pub fn new(
filename: PathBuf,
config: Option<Config>
) -> Result<Editor<'a>, Error>

Create a new Editor struct from a file.

pub fn default(config: Option<Config>) -> Result<Editor<'a>, Error>

Create a new Editor struct with the default welcome screen.

pub fn run(&mut self) -> Result<(), Error>

Run the editor logic.

Returns only when the user has signalled they want to quit.

pub fn action(&mut self, action: &str) -> Result<(), Error>

Actual implementor of Action.

Necessary due to borrow checker’s interaction with disjoint struct fields accessed through methods.

pub fn quit(&mut self)

Set the quit flag.

pub fn push_context<C>(&mut self, context: C) where
C: 'a + Context,

Push a Context to the stack of contexts.

pub fn push_boxed_context(
&mut self,
context: Box<dyn Context + 'static, Global>
)

Push a boxed Context to the stack of contexts.

pub fn q_draw_cmd_line<const N: usize>(
&mut self,
text: [&str; N],
flags: CmdLineFlags
) -> Result<(), Error>

Queue the necessary Commands to draw the command line.

Flushes the Command only if flush is true. If save_cursor is true, saves the cursor before the draw. If reset_cursor is true, after the draw the cursor will be placed where it was previously. Otherwise, it will remain on the command line after the drawn text.

pub fn terminal(&mut self) -> &mut Terminal

Return a reference to the terminal.

pub fn key_binds(&self) -> &KeyBinds

Return a reference to the KeyBinds object.

pub fn options(&self) -> &Options

Return a reference to the Options object.

pub fn config(&self) -> &Config

Return a reference to the Config object.

pub fn push_command(&mut self, cmd: String)

Push a command to the command history stack.

pub fn command_stack(&self) -> &Vec<String, Global>

Return a reference to the command history stack.

pub fn on_current_window<F, R>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&mut Window, &mut Terminal) -> Result<R, Error>,

Execute a closure taking mutable Window and Terminal references on the current window.