Module fim::layout

A module for handling keyboard layouts.

Custom Layout Specification Parsing

In addition to the three built-in layouts Qwerty, Dvorak, and Colemak, fim offers the ability for users to create their own custom keyboard layouts through ‘custom layout specifications’ (‘layout spec’ or just ‘spec’ for short), and use them in fim.

A layout spec is its own file, and can have whatever extension you like, although .fiml is traditional.

A spec is composed of two parts: the name line and the layout pairs. The name line must be the first line of the layout spec, and has the syntax layout <name>, where <name> is the name of the custom layout defined in that file.

Each layout pair maps a printable ASCII character to another printable ASCII character, like so: <char_qwerty> => <char_layout>, where <char_qwerty> is the QWERTY character, and <char_layout> is the character you would like <char_qwerty> to map to. For example, if you wanted to map any press of the 1 key to a, for example, the layout pair would be 1 => a. A spec can have any number of layout pairs, and pairs later in the file override pairs earlier in the file (I don’t know why you would need that, but it is what it is).

If a QWERTY key is not mapped from in the spec, then whenever that key is pressed, you’ll get that QWERTY character back.

Structs

Colemak

Struct that represents the Colemak keyboard layout

CustomLayout

Struct that represents custom, user-defined keyboard layouts.

Dvorak

Struct that represents the Dvorak keyboard layout.

Qwerty

Struct that represents the QWERTY keyboard layout.

Traits

Layout

An interface for keyboard layouts.

Functions

deshift_qwerty

Inverse of shift_qwerty().

shift_qwerty

Maps an ASCII, QWERTY press into the ASCII character that would be created by pressing Shift and that press at the same time. See also: deshift_qwerty().