Struct fim::editor::Editor

pub struct Editor<'a> { /* fields omitted */ }

Struct that represents the fim editor.

Implementations

impl<'a> Editor<'a>

pub fn new(
filename: PathBuf,
config: Option<Config>
) -> Result<Editor<'a>, Error>

Create a new Editor struct from a file.

pub fn default(config: Option<Config>) -> Result<Editor<'a>, Error>

Create a new Editor struct with the default welcome screen.

pub fn run(&mut self) -> Result<(), Error>

Run the editor logic.

Returns only when the user has signalled they want to quit.

pub fn action(&mut self, action: &str) -> Result<(), Error>

Actual implementor of Action.

Necessary due to borrow checker’s interaction with disjoint struct fields accessed through methods.

pub fn quit(&mut self)

Set the quit flag.

pub fn push_context<C>(&mut self, context: C) where
C: 'a + Context,

Push a Context to the stack of contexts.

pub fn push_boxed_context(
&mut self,
context: Box<dyn Context + 'static, Global>
)

Push a boxed Context to the stack of contexts.

pub fn draw_cmd_line<const N: usize>(
&mut self,
text: [&str; N]
) -> Result<(), Error>

Draw the command line.

(This refers to the last line in the terminal that you type ed commands or “:q” into in vim. I have no idea what it’s really called, so I refer to it as the command line).

See also: Self::q_draw_cmd_line().

pub fn q_draw_cmd_line<const N: usize>(
&mut self,
text: [&str; N],
flush: bool
) -> Result<(), Error>

Queue the necessary Commands to draw the command line.

See also: Self::draw_cmd_line().

pub fn terminal(&mut self) -> &mut Terminal

Return a reference to the terminal.

pub fn key_binds(&self) -> &KeyBinds

Return a reference to the KeyBinds object.

pub fn options(&self) -> &Options

Return a reference to the Options object.

pub fn config(&self) -> &Config

Return a reference to the Config object.

pub fn push_command(&mut self, cmd: String)

Push a command to the command history stack.

pub fn command_stack(&self) -> &Vec<String, Global>

Return a reference to the command history stack.